ml_service/requirements.txt
numpy
torch
nltk==3.5
elasticsearch==8.17.0
kafka-python==2.0.2
redis==5.2.1